Created System to interact with items on the floor and attach them to the player model.
Added 2 blueprint interfaces for the player controller to talk to the Interactable Objects and for the Interactable Objects to send their stored static mesh to the player character.

You can test the functionality in the level there is a BP_Interactable in the level go up to it (You'll need to be close) and right click. It will then attach a cube to your player model.